move:CommentSection 파일 위치 components/Layout으로 이동
